🔧feat: change c-style enum to c++ strongly-typed enum.

+++ tinyDB.cc
@@ -13,7 +13,7 @@ void handleStatus(StatusCode& statusCode) {
       break;
     case StatusCode::kSuccessAndExit:
       std::cout << "Bye~" << std::endl;
-      exit(StatusCode::kSuccessAndExit);
+      exit(static_cast<int>(StatusCode::kSuccessAndExit));
       break;
     case StatusCode::kUnrecognizeMetaCommand:
       std::cerr << "Error: unrecognized meta command, " 
@@ -25,7 +25,7 @@ void handleStatus(StatusCode& statusCode) {
       break;
     default:
       std::cerr << "Fatal Error: Unknown error!" << std::endl;
-      exit(StatusCode::kUnknownError);
+      exit(static_cast<int>(StatusCode::kUnknownError));
       break;
   }
 }
